+++ packages/components/src/components/range-tick/range-tick.styles.ts
@@ -1,0 +1,34 @@
+import { css } from 'lit';
+
+export default css`
+  :host {
+    --tick-height: var(--syn-spacing-x-small);
+    --tick-label-top: 10px;
+  }
+
+  .tick {
+    color: var(--syn-input-icon-color);
+    font-family: var(--syn-font-family);
+    font-weight: var(--syn-font-weight-normal);
+    line-height: var(--syn-line-height-denser);
+    width: 1px;
+  }
+
+  .tick-line {
+    background: var(--syn-color-neutral-400);
+    height: var(--tick-height);
+  }
+
+  .tick--subdivision .tick-line{
+    height: calc(var(--tick-height) / 2);
+  }
+
+  .tick-label {
+    align-items: center;
+    display: flex;
+    justify-content: center;
+    margin-top: var(--tick-label-top);
+    text-align: center;
+    white-space: nowrap;
+  }
+`;